Move PreferencesPage to Preferences package
