+++ src/languages/en.js
@@ -488,12 +488,7 @@ export default {
         defaultPaymentMethod: 'Default',
     },
     preferencesPage: {
-        mostRecent: 'Most recent',
-        mostRecentModeDescription: 'This will display all chats by default, sorted by most recent, with pinned items at the top.',
-        focus: '#focus',
-        focusModeDescription: '#focus – This will only display unread and pinned chats, all sorted alphabetically.',
         receiveRelevantFeatureUpdatesAndExpensifyNews: 'Receive relevant feature updates and Expensify news',
-        priorityMode: 'Priority mode',
         language: 'Language',
         languages: {
             english: 'English',
@@ -501,7 +496,12 @@ export default {
         },
     },
     priorityModesPage: {
+        priorityMode: 'Priority mode',
         explainerText: 'Choose whether to show all chats by default sorted with most recent with pinned items at the top, or #focus on unread pinned items, sorted alphabetically.',  
+        mostRecent: 'Most recent',
+        mostRecentModeDescription: 'This will display all chats by default, sorted by most recent, with pinned items at the top.',
+        focus: '#focus',
+        focusModeDescription: '#focus – This will only display unread and pinned chats, all sorted alphabetically.',
     },
     signInPage: {
         expensifyDotCash: 'New Expensify',
